testing a new script structure
